+++ bin/descripter.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+echo '[phpspec] Describing a class'
+vendor/bin/phpspec --config=app/phpspec.yml.dist desc $1