+++ src/TempoSimple/Bundle/SpaghettiBundle/Command/GenerateDailyReportCommand.php
@@ -18,9 +18,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class GenerateDailyReportCommand extends ContainerAwareCommand
 {
-    /**
-     * {@inheritdoc}
-     */
+    /** {@inheritdoc} */
     protected function configure()
     {
         $this->setName('tempo-simple:generate:daily-report');
@@ -31,9 +29,7 @@ class GenerateDailyReportCommand extends ContainerAwareCommand
         );
     }
 
-    /**
-     * {@inheritdoc}
-     */
+    /** {@inheritdoc} */
     public function execute(InputInterface $input, OutputInterface $output)
     {
         $timeCardrepository = $this->getContainer()->get('tempo_simple_spaghetti.time_card_reporitory');